refactor: #85 공통 클래스 `global` 패키지로 통일
