+++ src/main/java/com/waguwagu/weat/domain/analysis/model/dto/IsMemberSubmitAnalysisSettingDtoResponseWrapper.java
@@ -1,8 +1,7 @@
 package com.waguwagu.weat.domain.analysis.model.dto;
 
-import com.waguwagu.weat.domain.common.dto.ResponseDTO;
+import com.waguwagu.weat.global.model.ResponseDTO;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
 public class IsMemberSubmitAnalysisSettingDtoResponseWrapper extends ResponseDTO<IsMemberSubmitAnalysisSettingDTO.Response> {
