+++ src/main/java/com/waguwagu/weat/domain/analysis/model/dto/SubmitAnalysisSettingDtoResponseWrapper.java
@@ -1,8 +1,6 @@
 package com.waguwagu.weat.domain.analysis.model.dto;
 
-import com.waguwagu.weat.domain.common.dto.ResponseDTO;
-import lombok.Getter;
-import lombok.Setter;
+import com.waguwagu.weat.global.model.ResponseDTO;
 
 
 public class SubmitAnalysisSettingDtoResponseWrapper extends ResponseDTO<SubmitAnalysisSettingDTO.Response> {
