+++ src/main/java/com/waguwagu/weat/domain/group/model/dto/CreateGroupDtoResponseWrapper.java
@@ -1,6 +1,6 @@
 package com.waguwagu.weat.domain.group.model.dto;
 
-import com.waguwagu.weat.domain.common.dto.ResponseDTO;
+import com.waguwagu.weat.global.model.ResponseDTO;
 
 
 public class CreateGroupDtoResponseWrapper extends ResponseDTO<CreateGroupDTO.Response> {
