+++ src/main/java/com/waguwagu/weat/global/model/ResponseDTO.java
@@ -1,9 +1,7 @@
-package com.waguwagu.weat.domain.common.dto;
+package com.waguwagu.weat.global.model;
 
 import com.waguwagu.weat.global.exception.ErrorCode;
 import io.swagger.v3.oas.annotations.media.Schema;
-import org.springframework.http.HttpStatus;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
